Initial commit.
This initial commit includes the basic setup for REPL-based development
with ClojureScript and NodeJS.  The Lambda handler currently queries
DynamoDB for the tables that it needs, creating them if they are not
found.

The JavaScript AWS SDK's callback-centric approach has been adapted to
use `cljs.core.async` channels and the monadic abstrations provided by
[cats].  This, too, is an experiment and an effort to broaden my
horizons and hopefully improve my understanding of these techniques and
the benefits that they can bring to an error-prone, asynchronous world.

[cats]: https://github.com/funcool/cats